🚧 Move queries in SSR routes
