+++ packages/applications/ssr/src/app/laureats/[identifiant]/abandon/modele-reponse/route.ts
@@ -1,6 +1,11 @@
 import { mediator } from 'mediateur';
 
-import { Abandon } from '@potentiel-domain/laureat';
+import { Abandon, CahierDesCharges } from '@potentiel-domain/laureat';
+import { ConsulterCandidatureQuery } from '@potentiel-domain/candidature';
+import { ConsulterAppelOffreQuery, AppelOffre } from '@potentiel-domain/appel-offre';
+import { DateTime, IdentifiantProjet } from '@potentiel-domain/common';
+import { ConsulterUtilisateurQuery } from '@potentiel-domain/utilisateur';
+import { buildDocxDocument } from '@potentiel-infrastructure/document-builder';
 
 import { decodeParameter } from '@/utils/decodeParameter';
 import { IdentifiantParameter } from '@/utils/identifiantParameter';
@@ -10,17 +15,152 @@ export const GET = async (_: Request, { params: { identifiant } }: IdentifiantPa
   withUtilisateur(async (utilisateur) => {
     const identifiantProjet = decodeParameter(identifiant);
 
-    const modèleRéponse = await mediator.send<Abandon.GénérerModèleRéponseAbandonQuery>({
-      type: 'Document.Query.GénérerModèleRéponseAbandon',
+    const { nomComplet } = await mediator.send<ConsulterUtilisateurQuery>({
+      type: 'Utilisateur.Query.ConsulterUtilisateur',
       data: {
-        identifiantProjet,
-        identifiantUtilisateur: utilisateur.identifiantUtilisateur.email,
+        identifiantUtilisateur: utilisateur.identifiantUtilisateur.formatter(),
       },
     });
 
-    return new Response(modèleRéponse.content, {
+    const candidature = await mediator.send<ConsulterCandidatureQuery>({
+      type: 'Candidature.Query.ConsulterCandidature',
+      data: {
+        identifiantProjet,
+      },
+    });
+
+    const abandon = await mediator.send<Abandon.ConsulterAbandonQuery>({
+      type: 'Lauréat.Abandon.Query.ConsulterAbandon',
+      data: {
+        identifiantProjetValue: identifiantProjet,
+      },
+    });
+
+    const appelOffres = await mediator.send<ConsulterAppelOffreQuery>({
+      type: 'AppelOffre.Query.ConsulterAppelOffre',
+      data: { identifiantAppelOffre: candidature.appelOffre },
+    });
+
+    const { cahierDesChargesChoisi } =
+      await mediator.send<CahierDesCharges.ConsulterCahierDesChargesChoisiQuery>({
+        type: 'Lauréat.CahierDesCharges.Query.ConsulterCahierDesChargesChoisi',
+        data: { identifiantProjet },
+      });
+
+    const dispositionCDC = getCDCAbandonRefs({
+      appelOffres,
+      période: candidature.période,
+      cahierDesChargesChoisi,
+    });
+
+    const content = await buildDocxDocument({
+      type: 'abandon',
+      data: {
+        aprèsConfirmation: abandon.demande.confirmation?.confirméLe ? true : false,
+        adresseCandidat: candidature.candidat.adressePostale,
+        codePostalProjet: candidature.localité.codePostal,
+        communeProjet: candidature.localité.commune,
+        contenuParagrapheAbandon: dispositionCDC.dispositions,
+        dateConfirmation:
+          abandon.demande.confirmation?.confirméLe?.date.toLocaleDateString('fr-FR') || '',
+        dateDemande: abandon.demande.demandéLe.date.toLocaleDateString('fr-FR'),
+        dateDemandeConfirmation:
+          abandon.demande.confirmation?.demandéLe.date.toLocaleDateString('fr-FR') || '',
+        dateNotification: DateTime.convertirEnValueType(
+          candidature.dateDésignation,
+        ).date.toLocaleDateString('fr-FR'),
+        dreal: candidature.localité.région,
+        email: '',
+        familles: candidature.famille ? 'yes' : '',
+        ...getEdfType(candidature.localité.région),
+        justificationDemande: abandon.demande.raison,
+        nomCandidat: candidature.candidat.nom,
+        nomProjet: candidature.nom,
+        nomRepresentantLegal: candidature.candidat.représentantLégal,
+        puissance: candidature.puissance.toString(),
+        referenceParagrapheAbandon: dispositionCDC.référenceParagraphe,
+        refPotentiel: formatIdentifiantProjetForDocument(identifiantProjet),
+        status: abandon.statut.statut,
+        suiviPar: nomComplet || '',
+        suiviParEmail: appelOffres.dossierSuiviPar,
+        titreAppelOffre: appelOffres.title,
+        titreFamille: candidature.famille || '',
+        titrePeriode:
+          appelOffres.periodes.find((période) => période.id === candidature.période)?.title || '',
+        unitePuissance: appelOffres.unitePuissance,
+      },
+    });
+
+    return new Response(content, {
       headers: {
-        'content-type': modèleRéponse.format,
+        'content-type': 'application/vnd.openxmlformats-officedocument.wordprocessingml.document',
       },
     });
   });
+
+function getEdfType(region: string) {
+  if (!region) {
+    return {
+      isEDFOA: '',
+      isEDFSEI: '',
+      isEDM: '',
+    };
+  }
+
+  return {
+    isEDFOA: `${
+      !['Guadeloupe', 'Guyane', 'Martinique', 'Corse', 'La Réunion', 'Mayotte'].includes(region)
+        ? 'true'
+        : ''
+    }`,
+    isEDFSEI: `${
+      ['Guadeloupe', 'Guyane', 'Martinique', 'Corse', 'La Réunion'].includes(region) ? 'true' : ''
+    }`,
+    isEDM: `${region === 'Mayotte' ? 'true' : ''}`,
+  };
+}
+
+function getCDCAbandonRefs({
+  appelOffres,
+  période,
+  cahierDesChargesChoisi,
+}: {
+  appelOffres: AppelOffre;
+  période: string;
+  cahierDesChargesChoisi: string;
+}) {
+  const périodeDetails = appelOffres.periodes.find((periode) => periode.id === période);
+  const cdc = parseCahierDesChargesChoisi(cahierDesChargesChoisi);
+  const cahierDesChargesModifié = périodeDetails?.cahiersDesChargesModifiésDisponibles.find(
+    (c) => cdc.type === 'modifié' && c.paruLe === cdc.paruLe && c.alternatif === cdc.alternatif,
+  );
+
+  return {
+    référenceParagraphe: '!!!REFERENCE NON DISPONIBLE!!!',
+    dispositions: '!!!CONTENU NON DISPONIBLE!!!',
+    ...appelOffres.donnéesCourriersRéponse.texteEngagementRéalisationEtModalitésAbandon,
+    ...périodeDetails?.donnéesCourriersRéponse?.texteEngagementRéalisationEtModalitésAbandon,
+    ...(cahierDesChargesModifié &&
+      cahierDesChargesModifié.donnéesCourriersRéponse
+        ?.texteEngagementRéalisationEtModalitésAbandon),
+  };
+}
+
+const parseCahierDesChargesChoisi = (référence: string) => {
+  if (référence === 'initial') {
+    return { type: 'initial' };
+  }
+
+  return {
+    type: 'modifié',
+    paruLe: référence.replace('-alternatif', ''),
+    alternatif: référence.search('-alternatif') === -1 ? undefined : true,
+  };
+};
+
+const formatIdentifiantProjetForDocument = (identifiantProjet: string): string => {
+  const { appelOffre, période, famille, numéroCRE } =
+    IdentifiantProjet.convertirEnValueType(identifiantProjet);
+
+  return `${appelOffre}-P${période}${famille ? `-F${famille}` : ''}-${numéroCRE}`;
+};
