+++ packages/domain/lauréat/package.json
@@ -13,7 +13,6 @@
     "@potentiel-domain/core": "*",
     "@potentiel-domain/common": "*",
     "@potentiel-domain/document": "*",
-    "@potentiel-domain/modele-document": "*",
     "@potentiel-domain/candidature": "*",
     "@potentiel-domain/appel-offre": "*",
     "@potentiel-domain/utilisateur": "*"
