+++ packages/domain/lauréat/tsconfig.json
@@ -18,9 +18,6 @@
       "path": "../document"
     },
     {
-      "path": "../modele-document"
-    },
-    {
       "path": "../candidature"
     },
     {
