--- packages/domain/modele-document/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "@potentiel-domain/modele-document",
-  "version": "0.0.0",
-  "description": "Domain pour les modèles de document",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "private": true,
-  "scripts": {
-    "build": "tsc"
-  },
-  "dependencies": {
-    "@potentiel-domain/common": "*",
-    "@potentiel-domain/core": "*",
-    "@potentiel-libraries/monads": "*"
-  },
-  "devDependencies": {
-    "@potentiel-config/tsconfig": "*"
-  }
-}
