--- packages/domain/modele-document/src/index.ts
@@ -1,2 +1,0 @@
-// Port
-export type { GénérerModèleDocumentPort, OptionsGénération } from './générerModèleDocument.port';
