--- packages/domain/modele-document/tsconfig.json
@@ -1,18 +1,0 @@
-{
-  "extends": "@potentiel-config/tsconfig/tsconfig.base.json",
-  "compilerOptions": {
-    "outDir": "dist",
-    "rootDir": "src"
-  },
-  "references": [
-    {
-      "path": "../../libraries/monads"
-    },
-    {
-      "path": "../common"
-    },
-    {
-      "path": "../core"
-    }
-  ]
-}
