+++ packages/infrastructure/domain-adapters/src/index.ts
@@ -1,5 +1,4 @@
 export * from './projet';
 export * from './utilisateur';
 export * as DocumentAdapter from './document/document.adapter';
-export * as ModeleDocumentAdapter from './modele-document/modele-document.adapter';
 export * as CandidatureAdapter from './candidature/candidature.adapter';
