--- packages/infrastructure/domain-adapters/src/modele-document/modele-document.adapter.ts
@@ -1,4 +1,0 @@
-import { GénérerModèleDocumentPort } from '@potentiel-domain/modele-document';
-import { buildDocxDocument } from '@potentiel-infrastructure/document-builder';
-
-export const générerModèleDocument: GénérerModèleDocumentPort = buildDocxDocument;
